chore(core-state): use identifiers, fix getActiveDelegates trigger action
