+++ packages/core-kernel/src/ioc/identifiers.ts
@@ -72,6 +72,7 @@ export const Identifiers = {
     BlockProcessor: Symbol.for("Block<Processor>"),
     // State - @todo: better names that won't clash
     BlockState: Symbol.for("State<Block>"),
+    RoundState: Symbol.for("State<Round>"),
     StateBlockStore: Symbol.for("State<BlockStore>"),
     StateStore: Symbol.for("State<StateStore>"),
     StateBuilder: Symbol.for("State<StateBuilder>"),
